add substack (#37)
This pull request includes changes to add a newsletter subscription
feature to the website and some minor adjustments to the layout. The
most important changes include the addition of a newsletter component,
its integration into the homepage, and a small layout adjustment in the
base layout file.

Newsletter feature addition:

*
[`src/components/newsletter.astro`](diffhunk://#diff-35426a8f75bfaf0762ccdb72a3d8a54fcf7c09f15abe6a720139c6068b25f1a3R1-R6):
Added a new newsletter component with an embedded Substack iframe for
newsletter subscriptions.
*
[`src/pages/index.astro`](diffhunk://#diff-95d291e9ce4c8739cc7e65ff7bf0838dd5294cf39ab787ba51a42d08fb2df663R8):
Imported the new `Newsletter` component and included it in the homepage
layout.
[[1]](diffhunk://#diff-95d291e9ce4c8739cc7e65ff7bf0838dd5294cf39ab787ba51a42d08fb2df663R8)
[[2]](diffhunk://#diff-95d291e9ce4c8739cc7e65ff7bf0838dd5294cf39ab787ba51a42d08fb2df663R27)

Layout adjustment:

*
[`src/layouts/base.astro`](diffhunk://#diff-a259877e4dcc0824f2e30fab733274c0c1c0f5f9f7e36b9b7450d48bfddaa303R113):
Added a new line in the base layout file to improve readability and
structure.